refactoring withAbsoluteNode, menu option, withDesignFromLibrary
